+++ PPE3GSB/PPE3GSB/PPE3GSB.csproj
@@ -56,11 +56,11 @@
     <Compile Include="frmBase.Designer.cs">
       <DependentUpon>frmBase.cs</DependentUpon>
     </Compile>
-    <Compile Include="frmMenu.cs">
+    <Compile Include="frmMenuGestion.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="frmMenu.Designer.cs">
-      <DependentUpon>frmMenu.cs</DependentUpon>
+    <Compile Include="frmMenuGestion.Designer.cs">
+      <DependentUpon>frmMenuGestion.cs</DependentUpon>
     </Compile>
     <Compile Include="modFrais.cs">
       <SubType>Form</SubType>
@@ -94,14 +94,17 @@
     <Compile Include="frmCptRendu.Designer.cs">
       <DependentUpon>frmCptRendu.cs</DependentUpon>
     </Compile>
-    <Compile Include="rfmVisiteur.cs">
+    <Compile Include="frmVisiteur.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="rfmVisiteur.Designer.cs">
-      <DependentUpon>rfmVisiteur.cs</DependentUpon>
+    <Compile Include="frmVisiteur.Designer.cs">
+      <DependentUpon>frmVisiteur.cs</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="frmMenu.resx">
-      <DependentUpon>frmMenu.cs</DependentUpon>
+    <EmbeddedResource Include="frmMenuGestion.resx">
+      <DependentUpon>frmMenuGestion.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="modFrais.resx">
+      <DependentUpon>modFrais.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="remboursementFrais.resx">
       <DependentUpon>remboursementFrais.cs</DependentUpon>
@@ -143,8 +146,8 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <EmbeddedResource Include="rfmVisiteur.resx">
-      <DependentUpon>rfmVisiteur.cs</DependentUpon>
+    <EmbeddedResource Include="frmVisiteur.resx">
+      <DependentUpon>frmVisiteur.cs</DependentUpon>
     </EmbeddedResource>
     <EntityDeploy Include="mdlGSB.edmx">
       <Generator>EntityModelCodeGenerator</Generator>
