+++ PPE3GSB/PPE3GSB/frmContainer.cs
@@ -23,5 +23,10 @@ namespace PPE3GSB
             frapport.MdiParent = this;
             frapport.Show();
         }
+
+        private void frmContainer_Load(object sender, EventArgs e)
+        {
+
+        }
     }
 }
