--- PPE3GSB/PPE3GSB/frmMenu.cs
@@ -1,31 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace PPE3GSB
-{
-    public partial class frmMenu : Form
-    {
-        public frmMenu()
-        {
-            InitializeComponent();
-        }
-
-        private void btn2_Click(object sender, EventArgs e)
-        {
-            Form frmVisiteur = new Form();
-            frmVisiteur.Show();
-        }
-
-        private void button3_Click(object sender, EventArgs e)
-        {
-            Application.Exit();
-        }
-    }
-}
