+++ PPE3GSB/PPE3GSB/remboursementFrais.cs
@@ -28,5 +28,10 @@ namespace PPE3GSB
             modFrais frm = new modFrais();
             frm.Show();
         }
+
+        private void remboursementFrais_Load(object sender, EventArgs e)
+        {
+
+        }
     }
 }
